Implement test endpoints model and structure
