+++ tests/helpers.js
@@ -18,6 +18,24 @@ const agent = new https.Agent({rejectUnauthorized: false});
 
 const should = chai.should();
 
+export function setupMatrix(match) {
+  // this will tell the report
+  // to make an interop matrix with this suite
+  this.matrix = true;
+  this.report = true;
+  this.implemented = [...match.keys()];
+  this.rowLabel = 'Test Name';
+  this.columnLabel = 'Implementer';
+}
+
+export function addPerTestMetadata() {
+    // append test meta data to the it/test this.
+    this.currentTest.cell = {
+        columnId: this.currentTest.parent.title,
+        rowId: this.currentTest.title
+    };
+}
+
 // Javascript's default ISO timestamp contains milliseconds.
 // This lops off the MS part of the UTC RFC3339 TimeStamp and replaces
 // it with a terminal Z.
